coffee_shop_backend/requirements.txt
django 
djangorestframework
drf-yasg
psycopg2
djangorestframework-simplejwt
psycopg2-binary
# channels>=4.0.0
